Moves capitalized() and a() to engine/say.py, which is a more felicitous and cozy home for them.
