+++ engine/actor.py
@@ -3,7 +3,6 @@ from engine.globals import G
 from engine import item
 from engine import say
 from engine import tartarus
-from cronenbroguelike import util
 
 
 class _Statistic:
@@ -20,7 +19,7 @@ class _Statistic:
             message = f"{'in' if delta >= 0 else 'de'}creased by {abs(delta)}"
         else:
             message = "does not change"
-        say.insayne(f"{util.capitalized(self._NAME)} {message}.")
+        say.insayne(f"{say.capitalized(self._NAME)} {message}.")
 
     def _modify(self, delta, do_log=True):
         self._static_value += delta
@@ -79,7 +78,7 @@ class _VariableStatistic(_Statistic):
 
     def _log_current(self, delta):
         say.insayne(
-            f"{util.capitalized(self._NAME)} "
+            f"{say.capitalized(self._NAME)} "
             f"{'restored' if delta >= 0 else 'damaged'} by {abs(delta)}."
         )
 
@@ -117,7 +116,7 @@ class Insanity(_VariableStatistic):
 
     def _log_current(self, delta):
         say.insayne(
-            f"{util.capitalized(self._NAME)} "
+            f"{say.capitalized(self._NAME)} "
             f"{'increased' if delta >= 0 else 'assuaged'} by {abs(delta)}."
         )
